requirements.txt
pyserial>=3.5
Pillow>=9.0
numpy>=1.21
svgpathtools>=1.6.0
